updated numpy data type; was defaulting to float and could not slice

+++ skge/base.py
@@ -128,7 +128,7 @@ class StochasticTrainer(object):
     def _optim(self, xys):
         idx = np.arange(len(xys))
         self.batch_size = np.ceil(len(xys) / self.nbatches)
-        batch_idx = np.arange(self.batch_size, len(xys), self.batch_size)
+        batch_idx = np.arange(self.batch_size, len(xys), self.batch_size, dtype=np.int64)
 
         for self.epoch in range(1, self.max_epochs + 1):
             # shuffle training examples
